move interface RegExManagerTemplates manager->config
